move vnext solution from scratchpad repo
I thought I would be able to keep file history with changes made for
.net vnext so was working a different repo with intent of replacing the
files. But the solution structure changed in vs 2015 so the files won't
land in the same place anyway and version history will not be kept. the
new files are in the src/cloudscribe-core folder. once working the older
files will be removed