remove nuget folder
the nuspec files and batch files were used in my mvc5 prototype work.
now that vs 2015 builds every project as a nuget I think we will not
need that stuff. I kept a backup in case. I don't wan to confuse people
browsing the repo who might think we are using that stuff so I removed
it.