Update NewSiteViewModel.cs
In a Cloudscribe project using host names, the following validation error shows up in New Site and Site Info screens:
For Host name, only digits, numbers, - and _ allowed, no spaces allowed
so it will not let you enter something like: www.mydomain.com
I added the delimiter before the hyphen in the regular expression for folder names and host names. I added the period character as an allowed character in the regular expression for host name and added period to the descriptive error.