Update SiteManager.cs
In an implementation using domain based tenants, if the site admin went in and made a change to a tenant's setting (such as changing the theme or site name), if a user was on the tenant site, they would not see the changes made because the old settings were "stuck" in cache. The original implementation in the SiteManager Update function was getting the host name from the current request context rather than the from that of the passed in site ID. Also, the change accommodates the scenario where the tenant site has additional host mappings. I believe there would be an entry in cache with a key name based on the additional host name.